Updated API's for report

+++ src/apps/api/Models/Reports.cs
@@ -7,7 +7,7 @@ public class Reports
     public int userId { get; set; }
     public int systemId { get; set; }
     public int locationId { get; set; }
-    public int daylightHours { get; set; }
+    public float daylightHours { get; set; }
     public string? image { get; set; }
     public DateTime dateCreated { get; set; }
 }
